renamed gp to genetic in order to match gplearn naming convention
