+++ cpp/CMakeLists.txt
@@ -398,7 +398,7 @@ if(BUILD_CUML_CPP_LIBRARY)
     src/fil/fil.cu
     src/fil/infer.cu
     src/glm/glm.cu
-    src/gp/node.cu
+    src/genetic/node.cu
     src/holtwinters/holtwinters.cu
     src/kmeans/kmeans.cu
     src/knn/knn.cu
